backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Data validation and serialization
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP client and async support
httpx==0.25.2
aiofiles==23.2.1

# Background tasks (optional)
celery==5.3.4
redis==5.0.1

# Environment management
python-dotenv==1.0.0

# Web scraping and parsing
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10
